requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
SQLAlchemy==2.0.32
pymysql==1.1.1
cryptography==43.0.3
redis==5.0.7
rq==1.16.2
rq-scheduler==0.13.1
pandas==2.2.2
numpy==1.26.4
mplfinance==0.12.10b0
matplotlib==3.9.0
requests==2.32.3
python-dotenv==1.0.1
yfinance==0.2.65
vnstock==3.2.6
python-dateutil==2.8.2
pytz==2023.3
lunardate==0.2.2
PyYAML

